requirements.txt
absl-py==0.11.0
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
dicom2nifti==2.2.12
dm-tree==0.1.6
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.6.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
MedPy==0.4.0
mpunet @ file:///homes/pmcd/Documents/MultiPlanarUNet
nibabel==3.2.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip==21.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.16
protobuf==3.15.2
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.1.2
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.5.4
setuptools==53.0.0
SimpleITK==2.0.2
six==1.15.0
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.2
tensorflow-probability==0.11.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==4.3.3
typeguard==2.11.1
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
